pyproject.toml
# Copyright 2025 National Oceanography Centre
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#  http://www.apache.org/licenses/LICENSE-2.0
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "mamma-mia"
version = "0.3.0"
authors = [
  { name="thopri", email="thopri@noc.ac.uk" },
]
description = ""
readme = "README.md"
requires-python = "==3.11.*"
dependencies = ["loguru==0.7.*",
                "pyinterp==2024.6.*",
                "cattrs==24.1.*",
                "xarray==2024.7.*",
                "pyyaml==6.0.*",
                "zarr==2.18.*",
                "plotly==5.24.*",
                "intake==2.0.*",
                "intake-xarray==0.7.*",
                "copernicusmarine==2.1.*",
                "s3fs==2024.6.*",
                "blosc==1.11.*",
                "gsw==3.6.*",
                "xesmf@git+https://github.com/pangeo-data/xesmf.git",
                "glidersim@git+https://github.com/NOC-MDP/glidersim",
                "latlon@git+https://github.com/NOC-MDP/latlon",
                "GliderNetCDF@git+https://github.com/NOC-MDP/GliderNetCDF"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "Apache2.0"
license-files = ["LICENSE"]

[project.urls]
Homepage = "https://github.com/NOC-MDP/MammaMia"
Issues = "https://github.com/NOC-MDP/MammaMia/issues"

[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
mamma_mia = ["*.json"]